Split binary functions from math.cc
