+++ chainerx_cc/chainerx/cuda/CMakeLists.txt
@@ -26,6 +26,7 @@ CUDA_ADD_LIBRARY(chainerx_cuda STATIC
     cuda_device/activation.cu
     cuda_device/arithmetic.cu
     cuda_device/batch_norm.cc
+    cuda_device/binary.cu
     cuda_device/conv.cc
     cuda_device/copy.cu
     cuda_device/comparison.cu
