+++ chainerx_cc/chainerx/native/CMakeLists.txt
@@ -16,6 +16,7 @@ add_library(chainerx_native STATIC
     native_device/activation.cc
     native_device/arithmetic.cc
     native_device/batch_norm.cc
+    native_device/binary.cc
     native_device/conv.cc
     native_device/copy.cc
     native_device/comparison.cc
