+++ chainerx_cc/chainerx/routines/CMakeLists.txt
@@ -1,4 +1,5 @@
 add_library(chainerx_routines STATIC
+    binary.cc
     connection.cc
     creation.cc
     indexing.cc
@@ -14,6 +15,7 @@ add_library(chainerx_routines STATIC
 )
 
 install(FILES
+    binary.h
     connection.h
     creation.h
     indexing.h
