+++ chainerx_cc/chainerx/routines/math.h
@@ -149,29 +149,4 @@ Array Fabs(const Array& x);
 
 Array Sign(const Array& x);
 
-namespace internal {
-
-void IBitwiseAnd(const Array& x1, const Array& x2);
-void IBitwiseAnd(const Array& x1, Scalar x2);
-
-void IBitwiseOr(const Array& x1, const Array& x2);
-void IBitwiseOr(const Array& x1, Scalar x2);
-
-void IBitwiseXor(const Array& x1, const Array& x2);
-void IBitwiseXor(const Array& x1, Scalar x2);
-
-}  // namespace internal
-
-Array BitwiseAnd(const Array& x1, const Array& x2);
-Array BitwiseAnd(const Array& x1, Scalar x2);
-Array BitwiseAnd(Scalar x1, const Array& x2);
-
-Array BitwiseOr(const Array& x1, const Array& x2);
-Array BitwiseOr(const Array& x1, Scalar x2);
-Array BitwiseOr(Scalar x1, const Array& x2);
-
-Array BitwiseXor(const Array& x1, const Array& x2);
-Array BitwiseXor(const Array& x1, Scalar x2);
-Array BitwiseXor(Scalar x1, const Array& x2);
-
 }  // namespace chainerx
